+++ CHANGELOG.md
@@ -10,10 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Converted to a source generator
+- Renamed Trace to be an overload of WithTrackingName
+- Renumbered the GeneratorStage Enum values
 
 ### Added
 
 - GeneratorStage Enum and Descriptions for use with Lightweight Tracing
+- CancellationToken.ThrowIfCancellationRequested overload
+- Events (and counters) now support id & instance value
 
 ## [0.0.3-alpha] - 2025-07-19